+++ src/Venue.ts
@@ -7,6 +7,7 @@ import { OperationManager } from './OperationManager';
 import { WorkspaceManager } from './WorkspaceManager';
 import { UCANManager } from './UCANManager';
 import { SecretManager } from './SecretManager';
+import { SkillManager } from './SkillManager';
 import { Asset } from './Asset';
 import { fetchStreamWithError, fetchWithError } from './Utils';
 import { Auth, NoAuth } from './Credentials';
@@ -123,6 +124,7 @@ export class Venue implements VenueInterface {
   private _workspace?: WorkspaceManager;
   private _ucan?: UCANManager;
   private _secrets?: SecretManager;
+  private _skills?: SkillManager;
 
   get adapters(): AdapterManager { return this._adapters ??= new AdapterManager(this); }
   get agents(): AgentManager { return this._agents ??= new AgentManager(this); }
@@ -132,6 +134,7 @@ export class Venue implements VenueInterface {
   get workspace(): WorkspaceManager { return this._workspace ??= new WorkspaceManager(this); }
   get ucan(): UCANManager { return this._ucan ??= new UCANManager(this); }
   get secrets(): SecretManager { return this._secrets ??= new SecretManager(this); }
+  get skills(): SkillManager { return this._skills ??= new SkillManager(this); }
 
   constructor(options: VenueOptions = {}) {
     this.baseUrl = options.baseUrl || '';